+++ HISTORY.rst
@@ -4,6 +4,9 @@ Release History
 2.0.0 (2015-??-??)
 ------------------
 
+* Implemented PlatformIO CLI 2.0: "platform" related commands have been
+  moved to ``platformio platforms`` subcommand
+  (`issue #158 <https://github.com/platformio/platformio/issues/158>`_)
 * Added global ``-f, --force`` option which will force to accept any
   confirmation prompts (`issue #152 <https://github.com/platformio/platformio/issues/152>`_)
 * Allowed to choose which library to update